+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 [unreleased]
+- Fix: Sometimes the machine doesn't response to the initial machine "model" query. Attempt to query the model periodically until it's determined
+- Fix: Fixed single axis z probing.
 
 
 [0.10.0]